+++ src/utilities/index.ts
@@ -1,3 +1,8 @@
+import storage from '@react-native-firebase/storage';
+import { nanoid } from '@reduxjs/toolkit';
+
+import { MediaSource } from 'src/api';
+
 export * from './alert';
 
 // For now, this is okay for English localisations
@@ -52,4 +57,24 @@ export function generateRandomNumberBetween(min: number, max: number) {
  */
 export function getWordCount(input: string): number {
   return (input.trim().match(/\S+/g) ?? []).length;
+}
+
+type GenerateStoragePathConfig = { filename: string; isVideo: boolean };
+
+export function uploadFileToFirebase(
+  source: MediaSource,
+  generateStoragePath: (config: GenerateStoragePathConfig) => string,
+) {
+  const localFilePath = source.path ?? source.url;
+
+  const fileId = nanoid();
+  const isVideo = source.mime.includes('video');
+  const extension = isVideo ? 'mp4' : 'jpg';
+  const filename = `${fileId}.${extension}`;
+  const storagePath = generateStoragePath({ filename, isVideo });
+
+  const reference = storage().ref(storagePath);
+  const uploadTask = reference.putFile(localFilePath);
+
+  return [filename, uploadTask, reference] as const;
 }